Adding rocket vanilla scripts

+++ setup.sh
@@ -2,23 +2,39 @@
 
 mkdir -p ~/.config/scripts/
 
-cp evernote/evernote-things-sync.applescript ~/.config/scripts/evernote-things-sync.applescript
+uname_out="$(uname -s)"
+case "${uname_out}" in
+        Linux*) machine=Linux;;
+        Darwin*) machine=Mac;;
+        CYGWIN*) machine=Cygwin;;
+        MINGW*) machine=MinGw;;
+        *) machine="UNKNOWN::${uname_out}"
+esac
 
-./dev/install-nodejs.sh
-./dev/install-vim.sh
-./dev/setup-brew.sh
+if [ $machine = "Mac" ]; then
+        cp evernote/evernote-things-sync.applescript ~/.config/scripts/evernote-things-sync.applescript
 
-./mac/setup-mac.sh
+        ./dev/install-nodejs.sh
+        ./dev/install-vim.sh
+        ./dev/setup-brew.sh
 
-#write out current crontab
-crontab -l > evernote-cron
-#echo new cron into cron file
-echo "* * * * *  osascript ~/.config/scripts/evernote-things-sync.applescript" >> evernote-cron
-#install new cron file
-crontab evernote-cron
-rm evernote-cron
+        ./mac/setup-mac.sh
 
-# End
-killall Dock
-killall Finder
-sudo reboot
+        # TODO:
+        # Setup Fluid
+        # Setup Lens App
+        # Setup RescueTime
+
+        #write out current crontab
+        crontab -l > evernote-cron
+        #echo new cron into cron file
+        echo "* * * * *  osascript ~/.config/scripts/evernote-things-sync.applescript" >> evernote-cron
+        #install new cron file
+        crontab evernote-cron
+        rm evernote-cron
+
+        # End
+        killall Dock
+        killall Finder
+        sudo reboot
+fi
